models/mtcnn_project/pyproject.toml
[project]
name = "mtcnn-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]
